# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification - Diputados Paraguay
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Legislador {
  id: string;
  nombre: string;
  apellido: string;
  partido: string;
  bancada: string;
  departamento: string;
  cargo: string;
  periodo: string;
  foto?: string | null;
  email?: string | null;
  bio?: string | null;
  comisiones?: string[];
}

export interface LegisladoresResponse {
  data: Legislador[];
  total: number;
  page: number;
  totalPages: number;
}

export interface Comision {
  id: string;
  nombre: string;
  tipo: string;
  camara: string;
  presidente?: string | null;
  vicepresidente?: string | null;
  email?: string | null;
  miembros: string[];
}

export interface ComisionesResponse {
  data: Comision[];
  total: number;
}

export interface Sesion {
  id: string;
  fecha: string;
  horaInicio: string;
  horaFin?: string | null;
  tipo: string;
  estado: string;
  periodo: string;
  descripcion?: string | null;
  orden_del_dia?: string[];
}

export interface SesionesResponse {
  data: Sesion[];
  total: number;
  sesionEnVivo?: Sesion | null;
}

export type ProyectoHistorialItem = {
  fecha: string;
  evento: string;
  descripcion: string;
};

export interface Proyecto {
  id: string;
  numero: string;
  titulo: string;
  estado: string;
  etapa: string;
  fechaIngreso: string;
  iniciativa: string;
  comision?: string | null;
  descripcion?: string | null;
  appURL?: string | null;
  historial?: ProyectoHistorialItem[];
}

export interface ProyectosResponse {
  data: Proyecto[];
  total: number;
  page: number;
  totalPages: number;
}

export interface Ley {
  numero: string;
  titulo: string;
  fechaSancion: string;
  fechaPromulgacion?: string | null;
  proyecto?: string | null;
}

export interface LeyesResponse {
  data: Ley[];
  total: number;
  page: number;
  totalPages: number;
}

export interface DashboardData {
  totalLegisladores: number;
  totalComisiones: number;
  sesionesEsteMes: number;
  proyectosPendientes: number;
  leyesAprobadas: number;
  sesionEnVivo?: Sesion | null;
  proximasSesiones: Sesion[];
  ultimosProyectos: Proyecto[];
  ultimasLeyes: Ley[];
}

export type VotacionVotosItem = {
  legislador: string;
  partido: string;
  sentido: string;
};

export interface Votacion {
  id: string;
  titulo: string;
  descripcion: string;
  fecha: string;
  tipo: string;
  camara: string;
  favor: number;
  contra: number;
  abstenciones: number;
  ausentes: number;
  resultado: string;
  appURL?: string | null;
  proyectoId?: string | null;
  votos?: VotacionVotosItem[];
}

export interface VotacionesResponse {
  data: Votacion[];
  total: number;
}

export type SystemStatusRecursosItem = {
  recurso: string;
  ultimaActualizacion?: string | null;
  estado: string;
};

export interface SystemStatus {
  online: boolean;
  ultimaActualizacion?: string | null;
  fuente: string;
  recursos: SystemStatusRecursosItem[];
}

export type AIConsultRequestHistorialItem = {
  role: string;
  content: string;
};

export interface AIConsultRequest {
  pregunta: string;
  historial?: AIConsultRequestHistorialItem[];
}

export interface AIConsultResponse {
  respuesta: string;
  tipo: string;
  datos?: unknown | null;
  fuentes: string[];
}

export type GetLegisladoresParams = {
partido?: string;
departamento?: string;
search?: string;
page?: number;
limit?: number;
};

export type GetSesionesParams = {
estado?: string;
tipo?: string;
page?: number;
};

export type GetProyectosParams = {
estado?: string;
comision?: string;
search?: string;
page?: number;
limit?: number;
};

export type GetLeyesParams = {
anio?: number;
search?: string;
page?: number;
};

export type GetVotacionesParams = {
search?: string;
};