# lib/api-zod/src/generated/types/systemStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification - Diputados Paraguay
 * OpenAPI spec version: 0.1.0
 */
import type { SystemStatusRecursosItem } from './systemStatusRecursosItem';

export interface SystemStatus {
  online: boolean;
  ultimaActualizacion?: string | null;
  fuente: string;
  recursos: SystemStatusRecursosItem[];
}