# lib/api-zod/src/generated/types/votacion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification - Diputados Paraguay
 * OpenAPI spec version: 0.1.0
 */
import type { VotacionVotosItem } from './votacionVotosItem';

export interface Votacion {
  id: string;
  titulo: string;
  descripcion: string;
  fecha: string;
  tipo: string;
  camara: string;
  favor: number;
  contra: number;
  abstenciones: number;
  ausentes: number;
  resultado: string;
  appURL?: string | null;
  proyectoId?: string | null;
  votos?: VotacionVotosItem[];
}