# lib/api-zod/src/generated/types/votacionesResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification - Diputados Paraguay
 * OpenAPI spec version: 0.1.0
 */
import type { Votacion } from './votacion';

export interface VotacionesResponse {
  data: Votacion[];
  total: number;
}